LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-import-scope-end.tentative-expected.txt

Inside
Below limit

FAIL Scope-imported rule applies within scope, above limit assert_equals: expected "1" but got ""
PASS Scope-imported rule does not apply below limit
